final commit
the project can greatly improve and go above and beyond.

in the future to add to this program you can add more features to the card.
check to see if email is a valid emal and github is a valid link.

create a limit to the amount of people that can be adde to a team.